+++ parent_algorithm.h
@@ -15,9 +15,9 @@ public:
     #pragma pack(push, 4)
     struct Coordinate
     {
-        int16_t x;
-        int16_t y;
-        int16_t z;
+        int32_t x;
+        int32_t y;
+        int32_t z;
     };
     #pragma pack(pop)
     unsigned int counter;
